Make the Container class abstract
.. to enforce that subclasses implement myServiceType.
